requirements.txt
streamlit==1.32.0
numpy==1.26.4
tensorflow==2.14.0
scikit-learn==1.3.2
pyserial==3.5
shap==0.44.0
pandas==2.2.1
plotly==5.20.0
scipy==1.12.0